refactor: [BE] 회원가입 컨트롤러를 authController로 변경
